chore: move old scaffolding routes
